Implemented course scraper
WHAT:
--------------
Script to see if there is a free course seat in the course provided from users. Can also return if the course is valid for an initial check before adding the course into the automated system.

HOW:
--------------
Course information is transformed into a beautifulSoup parse tree through HTML web scraping. Tree is iterated through to observe the text contents of the web page