Started base script for database connectivity
Databaseutil will provide functions to query course requests and add new entries to Azure Cosmos DB. The data for the requests are instantiated as CourseRequest objects to be consumed by the rest of the application.

Environmental variables URI and KEY are from the container and set on Heroku. These must be set to connect to the DB.